shards/Ariel_Project/Scripts/Library/gardening-bot-main/gardening-bot-main/requirements.txt
opencv-python
python3_xlib
pyautogui
python-dotenv
firebase-admin